Database search implementation in tfhe. Supports 32 database entries atm.

+++ examples/server_tfhe.c
@@ -56,7 +56,7 @@ equal(LweSample* result, const LweSample* a, const LweSample* b, const int bits,
 	bootsCONSTANT(&result[ri], 1, ck);
 
 	// compare all bits of a and b
-	for (uint8_t i = 0; i < bits; i++)
+	for (uint32_t i = 0; i < bits; i++)
 		equal_bits(&result[ri], &a[i], &b[i], tmp, ck);
 
 	delete_gate_bootstrapping_ciphertext(tmp);
@@ -75,23 +75,23 @@ eval(EVAL_FUNC_PTR func, const int bits)
 	LweSample* ciphertext[2];
 	LweSample* result = new_gate_bootstrapping_ciphertext_array(bits, params);
 
-	for (uint8_t i = 0; i < 2; i++)
+	for (uint32_t i = 0; i < 2; i++)
 		ciphertext[i] = new_gate_bootstrapping_ciphertext_array(bits, params);
 
 	FILE* data = fopen(CIPHER_FILE, READMODE);
-	for (uint8_t i = 0; i < 2; i++) {
-		for (uint8_t j = 0; j < bits; j++)
+	for (uint32_t i = 0; i < 2; i++) {
+		for (uint32_t j = 0; j < bits; j++)
 			import_gate_bootstrapping_ciphertext_fromFile(data, &ciphertext[i][j], params);
 	} fclose(data);
 
 	func(result, ciphertext[0], ciphertext[1], bits, 0, ck);
 
 	FILE* output = fopen(DATA_FILE, WRITEMODE);
-	for (uint8_t i = 0; i < bits; i++)
+	for (uint32_t i = 0; i < bits; i++)
 		export_gate_bootstrapping_ciphertext_toFile(output, &result[i], params);
 	fclose(output);
 
-	for (uint8_t i = 0; i < 2; i++)
+	for (uint32_t i = 0; i < 2; i++)
 		delete_gate_bootstrapping_ciphertext_array(bits, ciphertext[i]);
 	delete_gate_bootstrapping_ciphertext_array(bits, result);
 	delete_gate_bootstrapping_cloud_keyset(ck);
@@ -112,30 +112,34 @@ database_search(EVAL_FUNC_PTR func, const char* search, const char* db, const in
 	LweSample* search_term = new_gate_bootstrapping_ciphertext_array(bits, params);
 	LweSample* result = new_gate_bootstrapping_ciphertext_array(db_size, params);
 
-	for (uint8_t i = 0; i < db_size; i++)
+	for (uint32_t i = 0; i < db_size; i++)
 		db_item[i] = new_gate_bootstrapping_ciphertext_array(bits, params);
 
+	// LOADING CIPHERTEXT FROM FILE
 	FILE* db_file = fopen(db, READMODE);
-	for (uint8_t i = 0; i < db_size; i++) {
-		for (uint8_t j = 0; j < bits; j++) {
+	for (uint32_t i = 0; i < db_size; i++) {
+		for (uint32_t j = 0; j < bits; j++) {
 			import_gate_bootstrapping_ciphertext_fromFile(db_file, &db_item[i][j], params);
 		}
 	} fclose(db_file);
 
 	FILE* search_file = fopen(search, READMODE);
-	for (uint8_t i = 0; i < bits; i++)
+	for (uint32_t i = 0; i < bits; i++)
 		import_gate_bootstrapping_ciphertext_fromFile(search_file, &search_term[i], params);
 	fclose(search_file);
 
-	for (uint8_t i = 0; i < db_size; i++)
+	// HOMOMORPHIC OPERATION
+	for (uint32_t i = 0; i < db_size; i++)
 		func(result, search_term, db_item[i], bits, i, ck);
 
+	// WRITING OUTPUT TO FILE
 	FILE* output = fopen(DATA_FILE, WRITEMODE);
-	for (uint8_t i = 0; i < db_size; i++)
+	for (uint32_t i = 0; i < db_size; i++)
 		export_gate_bootstrapping_ciphertext_toFile(output, &result[i], params);
 	fclose(output);
 
-	for (uint8_t i = 0; i < db_size; i++)
+	// CLEANUP
+	for (uint32_t i = 0; i < db_size; i++)
 		delete_gate_bootstrapping_ciphertext_array(bits, db_item[i]);
 	delete_gate_bootstrapping_ciphertext_array(bits, search_term);
 	delete_gate_bootstrapping_ciphertext_array(db_size, result);
